Update the device/model interface to pass the dispatch index down to the device.
At the device level, for example the shmem device, the user dispatch id is
actually a dispatch SET id.  This allows protocols to register multiple
device messages within a particular dispacth set.

More specifically, it allows task 0 to register dispatch 0 and then dispatch 1,
and task 1 to register dispatch 1 and then dispatch 0, and still maintains
a communication link.

DCMF did not have this notion of "dispatch sets" and therefore required the
protocol registrations to happen in the same order on all ranks.