Fix packet/message model interface for code compatibility with gcc and xl.
A template parameter is needed, not a static constant integer, for the size of
an array used as a method parameter. This will force the "user" (protocol code)
to invoke the model interface methods with the correct array size. Otherwise
the compiler will flag an error and exit.

Also, fixed the compile_time_assert code to check for stuff. C++ template code
must actually invoke the compile_time_assert method for the compiler to parse
the method.