Fix COMPILE_TIME_ASSERT macro so it works and moved asserts into constructors.
C++ template code does not seem to be even _parsed_ by the compiler if the
code is not referenced. This defeats the purpose of a compile-time assertion
macro that creates a syntax error when the expression evaluates to false.

The "work around" is to move the COMPILE_TIME_ASSERTION macros into other
methods that will be invoked - such as class constructors.  As an extra
precaution against "dead" code being left in by the compiler the macro
is now updated so that the assertion is wrapped in a 'if(0){}' block.