Remove the hints from the xmi_multicast_t. They never should have been there.
I temporarily added them to support the C api, but we need a better
solution.   Note that I added another temporary hack to support the
C api, but it doesn't affect the xmi interface this time.  Eventually we
need a real solution, but the low level C api is low priority.