Initial CAU device framework
This is the initial device support for CAU
It adds the framework for m-* calls to CAU
The CAU calls are not implemented yet, but a new CAU coll registration
class has been added that uses a local device and a shared memory
device.  This should allow for protocols to use the local device
and the global device.  Currently, it assumes the semantics
are active message, but we can add allsided support for those protocols
later.  CAU uses an active message style for broadcast.

The next steps are to add CCMI protocols to support this.