Modified pgas rt to use new interface.  Added a send and a sendpwq
This patch adds send and sendpwq to the native interface base class
The PGASRT has been converted to use the base send.  At some point
we may want to convert it to use sendpwq so it can pick up data types

This particular patch has only been tested on MPI and probably won't
compile for other platforms.  A subsequent patch will be added.

The upshot is that bg can start using the pgas rt collectives as well.
We will want to enable shared memory at some point for the pgas rt