Working version of CAU + CollShm for multisync
Some details of the work in the cau patches:

Fix quite a few bugs in our "mapping" code, which was never fully
or correctly ported to a "rank" based system.  It was based on a
torus originally, and some of the function was not correct, or had
terrible performance.

Build an "unexpected" barrier, so we can build geometries that
are not synchronized (important for openshmem)

Add topology routines (or fix existing ones) to construct our
"local", and "n-th master" topologies

Devise a strategy for allocating a classroute/cau id.  Implemented
the "single threaded" classroute allocation.  We still have more
to do to get the multithreaded version working.

We needed routines to build the mapping, this includes doing
reductions at init time, including a tricky "chicken-and-egg"
problem with topologies and running collectives that aren't
completely naive.

Needed to a platform neutral way to pass device specific information
to the hardware device at "m-*" issue time.  This is an acceptable
solution for PERCS, but it probably needs to be tested on BG
to make sure it's truly neutral.

Added a CCMI protocol to handle the "2 device" protocols.
This part is "easy".  The m-* are put in place to simplify the
protocol writer's job.  The hard part should be in the m-* and geometry.

Implemented the msync over lapi_cau