Simple Class Route Allocation
This is a simple test of a general class route allocation
implemented for the MPI adaptor.  This can be used to perform
a "split-phase" analyze, where, on geometry creation, an analyze
is performed for each set of registrations.  The registration
performs a local analyze, and inserts collectives into the list
on this local analyze.  It then contributes a 64 bit integer
to a reduction.  The reduction results can be used to do things
like allocate class routes or do some kind of analysis agreement
with the other nodes.  At the end of the allreduce operation,
the client is notified of the results, and a "global" analyze is
called to read the results and potentially insert more algorithms

split classroute allocation