Added "device info" void * to m-* c++ interface
This m-* is designed to carry device specific information from higher
level code to the device implementing m-*.  The immediate use for this
is to pass in a class route id.  Not all m-* require a class route,
so this dev info field is optional with M-* over other networks.

At some point, we may want to formalize the device info structure
programatically.  For now, the developers will need to know that
their protocols are being implemented on top of something that needs
a devinfo (classroute) as input.