Trac #245: Add a new node mapping interface: bool isLocal(size_t)
Usually all that is needed is to determine if a task is on the same node as the
calling task. Previously, this required a call to isPeer(size_t,size_t)
specifying the target task and something like '__global.mapping.task()' to get
the task identifier of the calling process. By adding this new interface
implementation may further optimize and eliminate several calculations in the
isPeer() code path.