Trac #59: Extension framework
An example extension, "EXT_torus_network", is defined and implemented in the
directory:
  api/extension/c/torus_network/

An unusable template extension, which can be used as a template to cut-n-paste
into a new extension, is defined and implemented in the directory:
  api/extension/c/template_extension/

Refer to the generated doxygen for more information on how to define,
implement, register, and use a PAMI extension.

This commit requires an associated buildtools update in order to configure the
source to compile specific extension(s) into the PAMI library. See the
following commit from the 'buildtools' git repository:
  d7ba5e7574438f02c052f40b287ad19724a50491