Fixed an issue with PAMI/LAPI contexts
The PAMI context "contains" a lapi context, so on callback delivery
from the lapi implementation of PAMI, the context delivered would be
"wrong" from the PAMI supplied context from createv.  To make these
equivalent, a kind of inheritance was used.  The lapi context
was moved to the front of the pami context structure so these would
be equivalent on callback delivery.  This required moving the storage
from the lapi context from the heap into pami context storage, and some
pointer fixups in the lapi library.