Trac #271: Add new collective shmem support to BGQ.
It is not enabled unless _COLLSHM is #defined.  It co-exists with the old
shmem protocol at this time.

The only protocols registered are barrier, bcast, and reduce on a
shmem-only geometry.  No composite protocol uses it yet.

The Platform.h shmem constants need to be tuned for BGQ.  With the current
values, I've run 32 processes per 4G node with BG_SHAREDMEMSIZE=128.
I have been unable to tune it for 64 processes so far.