First pass at 2 device allreduce protocol and CAU Multicombine
This is the general structure of the allreduce protocol.
Also checked in is a first implementation of CAU Multicombine
The allreduce protocol takes a "local" device and a "global"
device as template parameters.  The multicombine operation
is assumed to be a 2 sided operation.  The CAU implementation
of reduction is asynchronous at the root, so the multicombine
operation handles unexpected messages at the root internally
to multicombine.

In this pass of the code, the protocol is not implemented,
but the layout and template parameters are all in place.