2 Device Allreduce Protocol
This is a protocol for 2 device (local and global) allreduce
This protocol implements the chained pipelining over multicombine
and multicast to implement the allreduce protocol.

This is not yet tested, as device support has not been completely
implemented.  This checkin is so that others can get started if/when
the devices are ready to try.