CMVC 172515:  32 bit applications may fail to get shared memory segments
In some 32 bit applications with large data, it may not be possible to
allocate shared memory.  One scenario where this may occur is when
-bmaxdata is set to 0x80000000 at compile.  Another is when the application
has a lot of data compiled in.  The solution here is to still use shared
memory for 32 bit, as this has significant performance advantages, but to
disable shared memory when the app segment is large and the memory map fails.

We were not checking for error codes in the return of the memory region allocate.
If memory regions can not be created, we have to disable shared memory entirely
for the collective devices that use this memory.  Some flags were put in place
to do so.