CMVC 173086:  Assert in allreduce
This fixes a problem with CCMI allreduce.  The problem is that
the binomial allreduce algorithm has a synchronization step that
initializes a barrier and an allreduce and when the barrier completes, issues the
allreduce.  The problem is that the barrier can complete in the call to send.

We don't set up the allreduce lookup (in the geometry)  until after issuing the barrier
so when an the barrier completes inside the call to start() the allreduce,
the cached value is NULL.  This was not seen on BlueGene because during the send operation,
the BG internals do not advance the receive state.  On PERCS, the receive and send
state is advanced during a send operation.

Either way, this is a bug in the collectives implementation and CCMI needs to handle it.
The fix is to re-order some initialization to ensure that all objects are constructed,
can be looked up, and are in the right place at collective issue time.