Typed send performance improvement changes (D174426)
This fix changes how to determine whether a TypeCode object is contiguous or not.
It the code is compact and there are no empty spaces, then it will be considered
contiguous.

If the application is using the default copy data and the sender sets the
recv_copy hint in the dispatch data, then the TypeMachine is not being
run and, instead, the data is being directly sent over to the receiver.