Performance improvements for Pack/Unpack operation (T260, D175012)
PAMI uses the PAMI_TYPE_CONTIGUOUS tpye code for the Pack/Unpack
operation of the user data to the dgs pool, copying one byte at a
time.

In order to improve the performance, a new type code object is used,
PAMI_TYPE_CONTIG_MAX, which has a unit size of ULONG_MAX and an
atom size of 1.  The per-byte copying is, therefore, being replaced
with copying large chunck of data, therefore drastically reducing the
number of memcpy function calls and, consequently, increasing the
performance of the Pack/Unpack operation.

In addition to adding the new type code, the default copy operation
is changed to use memcpy directly, instead of the Math::copy method.