D173927:  Hang in geometry_create
We have an alternative unexpected geometry barrier class that should
be higher performing and not require the generic deivice.

This patch does a couple things:
1)  Removes the pgas UE barrier, which required delivering a callback
many times in a row using the generic device.  I suspect that for very
high iteration counts, somehow our data is corrupted or the callbacks
aren't being delivered correctly.
2)  Switches to use the CCMI default barriers, which use the standard
early arrival queue/post methods to determine if a barrier is early
arrival or not.

This should solve the geometry creation hang.  We still have some issue
where I expect our message was not being delivered correctly via
the device queue, so we probably should write a test case for that