D175161:  32 bit PESSL test cases fail on shared memory allocation
In the case where ESSL takes too much memory, it may not be possible
to get shared memory on all the tasks.  In this case, I had some debug
to catch when the shared memory manager could not be created.

The error path was to assert when shared memory was disabled, but this
debug was unintentionally left in.

We should disable shared memory collectives in the case where our
shared memory segment cannot be created