D174870:  Fix hang in geometry create barrier
The barrier unexpected registration was caching the geometry id.
This will not work when a geometry id is reused in a flow like this:

g=geometry_create(id)
geometry_destroy(g)
geometry_create(id)

This will cause the cache to return the old executor, and subsequently
be the wrong barrier to execute, and possibly a barrier that has no
reference.

The solution is to always grab the barrier from the id cache.