D174820:  CAU 4 byte packet striding to 8 bytes, pipelining fixes
This patch refactors the cau multicombine to be more careful about pipelining
to shared memory, handle unexpected/early arrival data, and workaround
CAU nuances, like the 4-byte to 8 byte stride issue that will show up
on >2 nodes for 4 byte datatypes.  In the code previous to this patch,
it was quite difficult to enable the pack/unpack operations, so this refactor
was necessary.

This patch should fix all known CAU reduce/allreduce correctness issues,
although the performance for rooted reduction may need some work since we are using
the reduce over allreduce protocol for CAU at this time.