D175432:  Fixed non-root broadcast data dependency and pipelining issues
There were some issues in early message arrival, pipelining to shared memory
and data dependency in the device code.

This patch cleans up some of those problems and makes the device/message interface
more consistent for CAU.  This should also pave the way for better datatype
support when we move to non-contiguous data.