D175351:  Ensure that shared memory addresses are not used across tasks
The shared memory device assumed that all tasks on a node had the same address
but this may not be the case.  When sharing addresses, we should be using offsets
from the start of the memory region, instead of actual addresses.

This patch also removes a hard coded initialization of the geometry world
string.  This requires an allreduce to initialize the world geometry, but should
allow for more robust geometry creation and optimization.