Performance improvements for non-contiguous data layouts (D175866)
The code changes target the data layout of the PAMI type machine.
There are two (2) optimization paths that work together to modify
the data layout such that, when running the type machine, the code
path stays in the inner most loop, thus avoiding unnecessary operations.

The first optimization path deals with the code instructions as the data
layout is built.  It looks at the previous code instructions and determines
whether it is possible to modify it and avoid adding additional instructions.
It works well for MPI, which uses a very iterative algorithm to build the
data layout.

The second optimization path deals with optimizing the data layout during
the completion of the type code.  It looks for sequences of identical copy
instructions and compresses them in a repetitive copy instruction.  It again
works well for MPI, because of the iterative nature of the data layout
generation.