Increase total number of dispatch IDs to 4096 (D176061)
We have 256 dispatch IDs: 128 for the user, 128 for internal.
Now we are running out of internal IDs between point-to-point
and collective so this change increases total number of 4096:
2048 for the user and 2048 for internal.

The rule is that point-to-point uses IDs from 2048 upward and
collective uses from 4095 downward.

This patch is corresponding to a LAPI fix, this changes the BG shared
memory segment allocations to take a template parameter for
dispatch table size