D176302:  Clear geometry id to geometry map at geometry destroy time.
This problem is caused when MPI re-uses geometry id's (which is expected
and often for applications like PESSL).  The problem here is that
the geometry id to geometry object map was not cleared for destroyed
geometry id's.

The destroyed geometry objects could be erroneously looked up by
early arrival packets from geometry create operations on remote
tasks.  This problem becomes more likely at large scale and
with asynchronous geometry creation.