Added an option to select pami(bg) shared memory or lapi shared memory
Performance team requested that we add an option to select between
lapi shared memory and the bg shared memory segment.  The default
is the bg shared memory segment for ccmi collectives.

The option is an environment variable:  PAMI_COLLECTIVES_SHM_DIRECT
If this is set to 0, then we will use lapi shared memory, if this
is set to 1, we will use bg shared memory.  The default is 1, as it
always has been in the past.