Consolidate Type machine copy instructions (T410)
Changed the optimization logic to use the previous instruction in a recursive fashion.  Each instruction has the code
of the previous instruction, which makes it easy to find the location of the previous instruction.  If the current
instruction needs to be consolidated with the previous instruction, it makes it easy to find the location of the
previous instruction, pop it out of the instruction stack, and add a consolidated instruction back on the instruction
stack.

Removed the optimized option in the TypeCode class constructor.  The class constructors are now as follows:
	TypeCode();
	TypeCode(void *code_addr, size_t code_size);
	TypeCode(size_t code_size, primitive_type_t primitive);