Trac #518: hardcode the 'split E dimension" cutoff for mu rget operations.
The cutoff is set to 2048 based on performance measurements. This means that
all three packet transfers will ''not'' be split for latency reasons. If the
data length will occupy four or more packets then the message is split with
half of the data sent on the E+ torus link and the remainder sent on the E-
torus link.