Issue 4262: fix insidious shmem pt2pt message queue bug.
When the shared memory fifo is full a message is created and queued for
processing when the fifo is no longer full. This message contains pointers
to the packet payload source (iovec arrays and pointers, contiguous source
data pointer, etc) and a pointer to the packet header source. The contract
with the application is that the payload source is not reusable by the
application until the callback is invoked, but the header source _is_
reusable by the application as soon as the post method returns.

The bug was that the shared memory message merely retained a pointer to the
original header source instead of copying the header source into the
'model state' storage.