Prevent shared memory be returned too earily (D179451, T455)
We cannot return the shared memory before all tasks finishes use it. Even
after the barrier of geometry_destroy tasks may still using the shared
memory to do cleanup work, e.g. BSR.

Instead of asking the leader to return the shared memory, we use the
first word in the region as flag to indicate if the region is safe to be
returned.  That flag will be set by the last guy who has done with the
region