Handle the primitive type value during type machine optimizations (T461)
PAMI primitive types are contiguous memory layouts that can be used
directly or as sub-types in other types.  When used directly, they are
well identified by the value of the primitive member of the TypeCode class.
When used as a sub-type in another type, the supra-type does not preserve
he primitive of the sub-type.

The type machine implementation must change to preserve the primitive
value of primitive types that are incorporated in user-defined data types,
according to the following rules:
1. if a single primitive type is added to a supra-type, then the supra-type
   preserves the primitive value of the primitive type;
2. if an simple copy instruction is added explicitly, then the primitive
   value of the supra-type is user defined;
3. if multiple types with different primitive values are added to the same
   supra-type, then the primitive value of the supra-type is user-defined;
4. shift instructions do not change the primitive value of the supra-type

The changes required for this ticket affect only the implementation of the
code instruction area of the type machine.