D179815:  Fix hang in asynchronous thread control structure allocations
In the optimized collective shared memory path used by BSR and CAU
for local portions of the collectives, it was possible to hit a
hang in shmem "channel" allocation and deallocation.

The coll shared memory device has optimizations not unlike flow control
to allow async style collective operations like bcast and reduce
to be both non-blocking and asyncronous (for example, root of a bcast
does not need to synchronize with the other tasks, and can finish a
collective immediately.

When a root task gets ahead of the other tasks in the geometry, it
will post a synchronizing operation after a certain number of collectives
The syncronizing operation cannot block to be pami compliant, so
a generic device "thread" is posted to the advance loop to handle
When there are no available shmem channels (root gets too far ahead),
it will post a monitor thread to handle the barrier syncronization,
and check for availability of the next round of threads (the other
collectives "catch up").

This monitor thread was posted to the generic device when the ref
counting on the messages hit zero and no threads are available.
The bug was that the reference counting was wrong, and caused the monitor
thread to not be invoked, thus not freeing resources for the thread.

The reference counting was only used for this purpose, so it has
been removed in this patch.  Instead, the device will post the monitor
thread only when out of resources, and no monitor thread is active.
Ref counting is just extra cycles and error prone.

This problem was timing sensitive and also sensitive to the control
structure circular queue layout.

Also found a couple places where the modulo arithmetic optimizations
(using bitshifts) were not completely disabled.  These have been flagged
with a comment that we can re-enable this performance optimization.