Initialize the original type member in TypeMachine to NULL (D180236)
The orig_type member of TypeMachine is not initialized in the TypeMachine constructor.
The optimized member of the TypeMachine class is used to determine whether the original
type has been modified, for historical reasons.

Now, the optimized member is used only for swapping the original type w/PAMI_TYPE_CONTIG_MAX.
This change removes the use of the optimized member and initializes the orig_type member
to NULL in the TypeMachine constructors.