Change the handling of the primitive information in PAMI's type machine (D180841)
The primitive information is currently maintained in the TypeCode object, as a
member of the TypeCode class, outside of the actual code executed by the type machine.

PAMI Put and Get operations that transfer the type code to the other party need
to preserve the primitive information during the transfering of the code.

If the primitive information becomes an attribute of the code itself, rather than
the TypeCode object, then it is recovered on the receiver's side in a consistent
manner.

This defect changes how the primitive information is handled such that it is
stored inside the code, as opposed to being stored in a member of the TypeCode
class.