D180974:  Fix problem with multi-context shared memory initialization
The low latency shared memory device used by BG/Q and CCMI for PE
also contains a research collective shared memory device.

This device has initialization problems on PE, but is not used.
The short term fix is to disable this until it can be enabled in
a more general way.