D181030:  Add templatized pad to maxloc/minloc structs for MPICH2
MPICH2 uses a pad between the short and int in the short int structure
based on compile time checks for the pad.

This fix ensures that we will find the correct offsets in the MPICH2
accumulate path to perform the max/minloc operation, and will not
destroy the pad value of the destination buffers.