Library code to dlopen and import Mellanox FCA libraries and symbols
This is the base code for importing Mellanox FCA hardware accelerated
collective functions.  The code sets up a class with function pointers
and dlopen/imports these function pointers from the libfca library from
mellanox.

It currently asserts and prints errors for missing symbols or libraries.
This code should be changed to detect the existence of the libraries
and selectively disable mellanox FCA if the library or symbols are not
present on the hardware.