trac628: create new 'is local task' extension.
The extension provides a platform-independent way of accessing PAMI
internals to determine if a task is local. This is done by exposing the
base address of an internal array, the byte stride used to index into
the array, and a bitmask to apply to the indexed byte. A task is
considered 'local' if the expression is non-zero.

This commit includes bgq and lapiunix implementations. The lapiunix
implementation using pami up-calls and needs to be optimized to use
lapiunix pami internals instead.