D183442:  LRZ Linpack PAMI_Wtime inaccuracy causes false noise
The code used for the x86 high resolution timers was based
on an old implementation from DCMF for RoadRunner.
This does not scale well to high core counts and potentially
heavily loaded systems.

The code replaced here is based on MPICH's use of high resolution
timers, and is calibrated more efficiently (no more spin to
calculate the clock hz).

The rdtsc instruction is replaced with a rdtscp syncronizing
instruction for more accurate timebase reads.

We considered the posix timers, but these are not suitable
for high resolution because they require a system call.