Issue 7604: Fix multi-threaded geometry create.
Geometry create (and GeometryOptimize) is asynchronous so a client can start
more than one create and wait for completion.   However there was static storage
for the GeometryOptimizer which was overwritten on each new create.  The
solution is to allocate storage for each create and free it in the cb_done.