Issue 8417: Fix world geometry (topology) rank order for alltoall
Previously, for performance, we switched from a range topology to a coordinate
topology because finding local and master subtologies is more efficient
using a coordinate topology.  This speeds up startup times.

However, we can't use just any topology in the world geometry.  It must be
equivalent to a range {0..size-1} topology and some coordinate topologies
might use a different rank ordering for different mappings.

So now only use the range topology by default.  However, for efficiency, pass
in any known coordinate and local topologies for the geometry to use to generate
it's local and master sub-topologies.

Signed-off-by: sssharka <sssharka@us.ibm.com>