Fix cpuid inline assembly to avoid register clobbering (D186683)
The original code attempts to save ebx into edi but it doesn't work
because x86_64 architecture by design zero-extends 32-bit value in
a 64-bit register. So the sequence below
   mov %%ebx, %%edi
   mov %%edi, %%ebx
essentially zeros out the high 32 bits of the 64-bit rbx register.

The new code is also simpler with compiler taking care of all the
output registers.